Rewrite streaming pipeline with proper H.264 handling
Overhaul the streaming pipeline with correct H.264 access unit
assembly and fMP4 segment packaging:

- Add H.264 bitstream parser (SPS/PPS decoding, NAL unit parsing)
  for extracting video dimensions from the stream
- Rewrite fMP4 segmenter with proper NAL-to-sample packaging,
  access unit boundaries, and correct box structure
- Improve WebRTC with RTP H.264 packetization and profile-level-id
  extraction from SPS
- Enhance FIFO reader with codec detection from ZoneMinder pipe
  filenames and reconnection improvements
- Refactor live coordinator for idempotent stop (204 vs 404) and
  cleaner session lifecycle management
- Add HLS storage pruning and segment count tracking

+++ src/repo/events.rs
@@ -158,7 +158,7 @@ pub async fn get_counts_by_monitor(
     #[derive(FromQueryResult)]
     struct CountResult {
         monitor_id: u32,
-        count: u64,
+        count: i64,
     }
 
     // Calculate time boundary
@@ -182,7 +182,7 @@ pub async fn get_counts_by_monitor(
 
     let counts = results
         .into_iter()
-        .map(|r| (r.monitor_id, r.count))
+        .map(|r| (r.monitor_id, r.count as u64))
         .collect();
 
     Ok(counts)
@@ -200,7 +200,7 @@ pub async fn get_counts_by_hour(
     #[derive(FromQueryResult)]
     struct CountResult {
         hour: String,
-        count: u64,
+        count: i64,
     }
 
     // Calculate time boundary
@@ -236,7 +236,7 @@ pub async fn get_counts_by_hour(
                         .unwrap()
                         .naive_utc()
                 });
-            (naive_date, r.count)
+            (naive_date, r.count as u64)
         })
         .collect();
